Move builder tests to AndroidEventBuilderTest
